removed example guy, moved example code to a comment in abstract class
