+++ memecity.engine/memecity.engine.vcxproj.filters
@@ -84,9 +84,6 @@
     <ClCompile Include="Updatable.cpp">
       <Filter>Source Files\Graphics</Filter>
     </ClCompile>
-    <ClCompile Include="ExampleGuy.cpp">
-      <Filter>Source Files</Filter>
-    </ClCompile>
   </ItemGroup>
   <ItemGroup>
     <ClInclude Include="GraphicsFacade.h">
@@ -136,9 +133,6 @@
     </ClInclude>
     <ClInclude Include="Updatable.h">
       <Filter>Header Files\Graphics</Filter>
-    </ClInclude>
-    <ClInclude Include="ExampleGuy.h">
-      <Filter>Header Files</Filter>
     </ClInclude>
   </ItemGroup>
 </Project>
